chatty unchecked bitstream report - includes item,collection,community info where available
