+++ dspace-api/src/main/java/org/dspace/checker/BitstreamInfo.java
@@ -18,9 +18,9 @@ import java.util.Date;
  * @author Jim Downing
  * @author Grace Carpenter
  * @author Nathan Sarr
- * 
+ *
  */
-public final class BitstreamInfo
+public final class BitstreamInfo extends CheckerInfo
 {
     /** deleted flag. */
     private boolean deleted;
@@ -47,13 +47,6 @@ public final class BitstreamInfo
     private Date processEndDate;
 
     /**
-     * Blanked off no-op default constructor.
-     */
-    private BitstreamInfo()
-    {
-    }
-
-    /**
      * Minimal constructor.
      * 
      * @param bid
@@ -61,6 +54,7 @@ public final class BitstreamInfo
      */
     public BitstreamInfo(int bid)
     {
+        super(bid);
         deleted = false;
 
         dspaceBitstream = new DSpaceBitstreamInfo(bid);
@@ -108,6 +102,7 @@ public final class BitstreamInfo
             String chksumAlgorthm, String chksum, String nm, Date procEndDate,
             boolean toBeProc, Date procStartDate)
     {
+        super(bitstrmId);
         dspaceBitstream = new DSpaceBitstreamInfo(del, storeNo, sz, bitstrmFmt,
                 bitstrmId, usrFmtDesc, intrnlId, src, chksumAlgorthm, chksum,
                 nm, "");
@@ -150,17 +145,6 @@ public final class BitstreamInfo
     }
 
     /**
-     * Set the store number.
-     * 
-     * @param storeNumber
-     *            the store number
-     */
-    public void setStoreNumber(int storeNumber)
-    {
-        dspaceBitstream.setStoreNumber(storeNumber);
-    }
-
-    /**
      * Get the size.
      * 
      * @return int
@@ -168,17 +152,6 @@ public final class BitstreamInfo
     public long getSize()
     {
         return dspaceBitstream.getSize();
-    }
-
-    /**
-     * Set the size.
-     * 
-     * @param size
-     *            the bitstream size
-     */
-    public void setSize(int size)
-    {
-        dspaceBitstream.setSize(size);
     }
 
     /**
@@ -191,16 +164,6 @@ public final class BitstreamInfo
         return dspaceBitstream.getBitstreamFormatId();
     }
 
-    /**
-     * Set the Bitstream Format id.
-     * 
-     * @param bitstrmFmt
-     *            id of the bitstream format
-     */
-    public void setBitstreamFormatId(String bitstrmFmt)
-    {
-        dspaceBitstream.setBitstreamFormatId(bitstrmFmt);
-    }
 
     /**
      * Get the Bitstream id.
@@ -223,17 +186,6 @@ public final class BitstreamInfo
     }
 
     /**
-     * Set the user format description.
-     * 
-     * @param userFormatDescription
-     *            the userFormatDescription
-     */
-    public void setUserFormatDescription(String userFormatDescription)
-    {
-        dspaceBitstream.setUserFormatDescription(userFormatDescription);
-    }
-
-    /**
      * Get the Internal Id.
      * 
      * @return String
@@ -241,17 +193,6 @@ public final class BitstreamInfo
     public String getInternalId()
     {
         return dspaceBitstream.getInternalId();
-    }
-
-    /**
-     * Set the Internal Id.
-     * 
-     * @param internalId
-     *            the DSpace internal sequence id for the bitstream.
-     */
-    public void setInternalId(String internalId)
-    {
-        dspaceBitstream.setInternalId(internalId);
     }
 
     /**
@@ -265,17 +206,6 @@ public final class BitstreamInfo
     }
 
     /**
-     * Set the source.
-     * 
-     * @param source
-     *            The bitstream source.
-     */
-    public void setSource(String source)
-    {
-        dspaceBitstream.setSource(source);
-    }
-
-    /**
      * Get the checksum algorithm.
      * 
      * @return String
@@ -283,17 +213,6 @@ public final class BitstreamInfo
     public String getChecksumAlgorithm()
     {
         return dspaceBitstream.getChecksumAlgorithm();
-    }
-
-    /**
-     * Set the checksum algorithm.
-     * 
-     * @param checksumAlgorithm
-     *            the algorithm used for checking this bitstream
-     */
-    public void setChecksumAlgorithm(String checksumAlgorithm)
-    {
-        dspaceBitstream.setChecksumAlgorithm(checksumAlgorithm);
     }
 
     /**
@@ -307,17 +226,6 @@ public final class BitstreamInfo
     }
 
     /**
-     * Set the checksum.
-     * 
-     * @param checksum
-     *            The last stored checksum for this bitstream.
-     */
-    public void setStoredChecksum(String checksum)
-    {
-        dspaceBitstream.setStoredChecksum(checksum);
-    }
-
-    /**
      * Get the name of the bitstream.
      * 
      * @return String
@@ -327,16 +235,6 @@ public final class BitstreamInfo
         return dspaceBitstream.getName();
     }
 
-    /**
-     * Set the name of the bitstream.
-     * 
-     * @param nm
-     *            The name of this bitstream.
-     */
-    public void setName(String nm)
-    {
-        dspaceBitstream.setName(nm);
-    }
 
     /**
      * calculatedChecksum accessor.
@@ -370,18 +268,7 @@ public final class BitstreamInfo
         return this.bitstreamFound;
     }
 
-    /**
-     * sets bitstreamFound.
-     * 
-     * @param found
-     *            value of bitstreamFound to set.
-     */
-    public void setBitstreamFound(boolean found)
-    {
-        this.bitstreamFound = found;
-    }
-
-    /**
+   /**
      * Identity entirely dependent upon <code>bitstreamId</code>.
      * 
      * @see java.lang.Object#equals(java.lang.Object)
@@ -420,8 +307,7 @@ public final class BitstreamInfo
      */
     public String toString()
     {
-        return new StringBuffer("ChecksumInformation for id ").append(
-                getBitstreamId()).toString();
+        return "ChecksumInformation for id "+ String.valueOf(getBitstreamId());
     }
 
     /**
